adk/totem_fashion/requirements.txt
# Project dependencies for the Totem Fashion Finder.
#
# This file declares the Python packages required to run the web API and the
# Google Agent Development Kit (ADK) integration.  Installing these
# dependencies with `pip install -r requirements.txt` will ensure the
# environment has FastAPI, Uvicorn, Pydantic for the API layer, and
# google-adk and google-generativeai for the Gemini agent.

fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
requests>=2.31.0

# Google Cloud / AI dependencies
google-adk>=0.3.0
google-generativeai>=0.6.0
google-cloud-aiplatform>=1.64.0
google-auth>=2.33.0